LayoutTests/fast/images/text-recognition/image-overlay-with-attachment-element-expected.txt

PASS transform is not 'none'
PASS successfullyParsed is true

TEST COMPLETE
